Conflicts:
	modules/ballerina-core/pom.xml
	modules/features/ballerina-core-feature/pom.xml
	pom.xml

+++ modules/ballerina-core/src/main/java/org/wso2/ballerina/core/nativeimpl/lang/json/AddDoubleToArray.java
@@ -39,7 +39,7 @@ import org.wso2.ballerina.core.nativeimpl.annotations.BallerinaFunction;
         functionName = "add",
         args = {@Argument(name = "json", type = TypeEnum.JSON),
                 @Argument(name = "jsonPath", type = TypeEnum.STRING),
-                @Argument(name = "value", type = TypeEnum.INT)},
+                @Argument(name = "value", type = TypeEnum.DOUBLE)},
         isPublic = true
 )
 @Component(
